+++ gdtoolkit/formatter/class_statement.py
@@ -8,7 +8,11 @@ from .context import Context, ExpressionContext
 from .block import format_block
 from .function_statement import format_func_statement
 from .enum import format_enum
-from .statement_utils import format_simple_statement
+from .statement_utils import (
+    format_simple_statement,
+    find_tree_among_children,
+    find_token_among_children,
+)
 from .var_statement import format_var_statement
 from .expression_to_str import expression_to_str
 from .expression import format_comma_separated_list, format_expression
@@ -270,95 +274,41 @@ def _format_func_statement(statement: Tree, context: Context) -> Outcome:
     return (formatted_lines, last_processed_line_no)
 
 
-# TODO: refactor that beast ^^
 def _format_func_header(statement: Tree, context: Context) -> Outcome:
+    return_type = find_token_among_children("TYPE", statement)
+    suffix_w_parent = ":" if return_type is None else f" -> {return_type.value}:"
+    parent_call = find_tree_among_children("parent_call", statement)
+    suffix_wo_parent = suffix_w_parent if parent_call is None else ""
     name_token = statement.children[0]
     name = name_token.value
-    func_args = (
-        statement.children[1]
-        if len(statement.children) > 1
-        and isinstance(statement.children[1], Tree)
-        and statement.children[1].data == "func_args"
-        else None
-    )
+    func_args = find_tree_among_children("func_args", statement)
     if func_args is not None:
         expression_context = ExpressionContext(
-            "func {}(".format(name), statement.line, ")", func_args.end_line
+            f"func {name}(",
+            statement.line,
+            f"){suffix_wo_parent}",
+            func_args.end_line,
         )
         formatted_lines = format_comma_separated_list(
             func_args.children, expression_context, context
         )
     else:
         formatted_lines = [
-            (name_token.line, "{}func {}()".format(context.indent_string, name))
+            (
+                name_token.line,
+                f"{context.indent_string}func {name}(){suffix_wo_parent}",
+            )
         ]
-    parent_call = (
-        statement.children[1]
-        if len(statement.children) > 1
-        and isinstance(statement.children[1], Tree)
-        and statement.children[1].data == "parent_call"
-        else None
-    )
-    parent_call = (
-        statement.children[2]
-        if len(statement.children) > 2
-        and isinstance(statement.children[2], Tree)
-        and statement.children[2].data == "parent_call"
-        else parent_call
-    )
     if parent_call is not None:
         last_line_no, last_line = formatted_lines[-1]
         expression_context = ExpressionContext(
-            "{}.(".format(last_line.strip()),
+            f"{last_line.strip()}.(",
             last_line_no,  # type: ignore
-            ")",
+            f"){suffix_w_parent}",
             parent_call.end_line,
         )
         elements = [e for e in parent_call.children[1:-1] if not is_any_comma(e)]
         formatted_lines = formatted_lines[:-1] + format_comma_separated_list(
             elements, expression_context, context
         )
-    return_type = (
-        statement.children[1]
-        if len(statement.children) > 1
-        and isinstance(statement.children[1], Token)
-        and statement.children[1].type == "TYPE"
-        else None
-    )
-    return_type = (
-        statement.children[2]
-        if len(statement.children) > 2
-        and isinstance(statement.children[2], Token)
-        and statement.children[2].type == "TYPE"
-        else return_type
-    )
-    return_type = (
-        statement.children[3]
-        if len(statement.children) > 3
-        and isinstance(statement.children[3], Token)
-        and statement.children[3].type == "TYPE"
-        else return_type
-    )
-    if return_type is not None:
-        last_line_no, last_line = formatted_lines[-1]
-        expression_context = ExpressionContext(
-            "{} -> ".format(last_line.strip()),
-            last_line_no,  # type: ignore
-            ":",
-            return_type.end_line,
-        )
-        formatted_lines = formatted_lines[:-1] + [
-            (
-                last_line_no,
-                "{}{} -> {}:".format(
-                    context.indent_string, last_line.strip(), return_type.value
-                ),
-            )
-        ]
-    else:
-        last_line_no, last_line = formatted_lines[-1]
-        formatted_lines = formatted_lines[:-1] + [
-            (last_line_no, "{}:".format(last_line))
-        ]
-    # import pdb;pdb.set_trace()
     return (formatted_lines, statement.end_line)