+++ tests/formatter/input-output-pairs/bug-148-complex-func-header.in.gd
@@ -1,0 +1,2 @@
+func decompress_lzw(code_stream_data: PoolByteArray, min_code_size: int, colors: PoolByteArray) -> PoolByteArray:
+	return code_stream_data