+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Fixed lack of trailing comma handler in some corner cases
  - Fixed lack of support for plus prefixed expressions
  - Fixed lack of support for `Vector2/3()` in `match` patterns
+ - Fixed formatter's corner case in `func()`
 
 ## [3.3.1] 2021-09-15